fix(TST-09): align fire animation state with time

+++ src/lincity/modules/firestation.hpp
@@ -97,7 +97,7 @@ public:
 
     int xs, ys, xe, ye;
     int daycount, covercount;
-    int anim;
+    unsigned long anim;
     bool animate_enable, active;
     int working_days, busy;
 };
